requirements.txt
requests~=2.32.3
psycopg2-binary==2.9.9
psycopg2==2.9.9
python-dotenv==1.0.1
nltk==3.8.1
spacy==3.7.4
spacy-legacy==3.0.12
spacy-loggers==1.0.5
google==3.0.0
google-cloud-storage==2.16.0